remove old blkalgn results
These results were captured with a buggy blkagln which shifted results
on json by 1, its not bad as you have to just imagine that 32k is 16k
but best to just not present those so remove it completely to not
confuse folks. There is already enough with the non double write
results.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>